+++ gameobjects/gameobject.h
@@ -2,7 +2,7 @@
 
 #include "colliderbox.h"
 #include "mesh.h"
-#include "audio.h"
+#include "sound.h"
 #include "utils.h"
 
 #include <glm/glm.hpp>
